+++ src/azure-cli/HISTORY.rst
@@ -12,6 +12,10 @@ Release History
 * add "user update" command
 * deprecate "--upn-or-object-id" from user related commands and introduce "--id"
 
+**SQL**
+
+* New Cmdlets for Management.Sql that supports Managed instance key and managed instance TDE protector management
+
 2.0.69
 ++++++